requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
beautifulsoup4==4.12.3
    # via -r requirements.in
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
cryptography==42.0.7
    # via pdfminer-six
distlib==0.3.8
    # via virtualenv
docstring-parser==0.16
    # via google-cloud-aiplatform
filelock==3.14.0
    # via virtualenv
google-api-core[grpc]==2.19.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-cloud-aiplatform==1.51.0
    # via -r requirements.in
google-cloud-bigquery==3.22.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.12.3
    # via google-cloud-aiplatform
google-cloud-storage==2.16.0
    # via
    #   -r requirements.in
    #   google-cloud-aiplatform
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.63.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-resource-manager
grpcio==1.63.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
identify==2.5.36
    # via pre-commit
idna==3.7
    # via requests
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via shapely
packaging==24.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
pdfminer-six==20231228
    # via pdfplumber
pdfplumber==0.11.0
    # via -r requirements.in
pillow==10.3.0
    # via pdfplumber
platformdirs==4.2.1
    # via virtualenv
pre-commit==3.7.1
    # via -r requirements.in
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via google-cloud-aiplatform
pydantic-core==2.18.2
    # via pydantic
pypdf2==3.0.1
    # via -r requirements.in
pypdfium2==4.30.0
    # via pdfplumber
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.9
    # via google-auth
shapely==2.0.4
    # via google-cloud-aiplatform
six==1.16.0
    # via python-dateutil
soupsieve==2.5
    # via beautifulsoup4
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
virtualenv==20.26.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
